backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymupdf>=1.23.0
chromadb>=0.4.22
httpx>=0.26.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
sse-starlette>=1.8.0
pillow>=10.2.0
numpy>=1.26.0
openai>=1.10.0
python-dotenv>=1.0.0
